requirements.txt
# 基础包
numpy==1.23.5
pandas==1.5.3

# PyTorch 相关（CUDA 12.1 版本）
# torch==2.0.1+cu121
# torchvision==0.15.2+cu121

# 数据处理与增强
Pillow==9.5.0
opencv-python==4.7.0.72

# 深度学习框架
pytorch-lightning==2.0.2

# 模型与工具
timm==0.6.13

# 评估指标
scikit-learn==1.2.2

# 其他工具
tqdm==4.65.0  # 进度条工具（可选）